+++ src/main/java/com/github/siboxd/fatturapa/model/DatiRiepilogoType.java
@@ -1,6 +1,6 @@
 package com.github.siboxd.fatturapa.model;
 
-import com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund.NaturaType;
+import com.github.siboxd.fatturapa.model.invoicebody.NaturaType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
