--- src/main/java/com/github/siboxd/fatturapa/model/TipoCessionePrestazioneType.java
@@ -1,15 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "TipoCessionePrestazioneType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum TipoCessionePrestazioneType {
-
-    SC,
-    PR,
-    AB,
-    AC
-
-}
