+++ src/main/java/com/github/siboxd/fatturapa/model/invoicebody/NaturaType.java
@@ -1,4 +1,4 @@
-package com.github.siboxd.fatturapa.model.invoicebody.general.pensionfund;
+package com.github.siboxd.fatturapa.model.invoicebody;
 
 import org.simpleframework.xml.Namespace;
 import org.simpleframework.xml.Root;
